add enums and hooks
