--- src/components/ComplexLookupField/configs/SearchResultsTableConfigs.ts
@@ -1,6 +1,0 @@
-import { authoritiesTableConfig } from './Authorities';
-
-export const SEARCH_RESULTS_TABLE_CONFIG: Record<string, SearchResultsTableConfig> = {
-  default: authoritiesTableConfig,
-  authorities: authoritiesTableConfig,
-};
