+++ src/components/ComplexLookupField/configs/index.ts
@@ -1,1 +1,6 @@
-export { SEARCH_RESULTS_TABLE_CONFIG } from './SearchResultsTableConfigs';
+import { authoritiesTableConfig } from './Authorities';
+
+export const SEARCH_RESULTS_TABLE_CONFIG: Record<string, SearchResultsTableConfig> = {
+  default: authoritiesTableConfig,
+  authorities: authoritiesTableConfig,
+};
